requirements.txt
# Core frameworks - Google ADK and LangGraph
google-genai>=1.0.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# Research APIs
arxiv>=2.1.0
semanticscholar>=0.8.0
requests>=2.31.0

# Vector store and embeddings
faiss-cpu>=1.8.0
numpy>=1.26.0

# Code formatting and quality
black>=24.0.0
isort>=5.13.0
autopep8>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Configuration and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
tenacity>=8.2.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Database for LangGraph checkpoints
aiosqlite>=0.19.0

# Optional: for enhanced logging
colorlog>=6.8.0

# FastAPI web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
